requirements.txt
oracledb
mysql.connector